[#49] Feat: TeamSelectSheetView 구현
